README.md
This project demonstrates the development of a hostel security system using an Arduino Uno and RFID reader with a beep alarm for incorrect ID and data collection. The system uses RFID tags and a reader to identify and authenticate authorized users. When an authorized user swipes their RFID tag on the reader, the Arduino Uno unlocks the door or gate, allowing the user to enter. If an unauthorized user swipes their RFID tag on the reader, the Arduino Uno sounds a beep alarm and denies access. The system also collects data on the RFID tags that are swiped, including the ID number and the date and time.